graphic display fix, starting lesioning

+++ replication-data/data-scripts/lesion.py
@@ -1,11 +1,15 @@
 from __future__ import division
 import sys
 from random import gauss
+import os.path
 
 GAUSSIAN_MEAN = 0
 
-def lesion_assocs (patient_file, lesion_file, map_sizes, num_reps, noise_amount):
+def lesion_assocs (patient_file, lesion_file, map_sizes, num_reps, noise_amount, overwrite):
 	
+	if os.path.isfile(lesion_file) and not overwrite:
+		print ("lesion_file {} already exists... not overwriting\n".format(lesion_file))
+		return
 	patient = open(patient_file, "r")
 	lesioned = open(lesion_file, "w")
 
@@ -38,7 +42,7 @@ def lesion_assocs (patient_file, lesion_file, map_sizes, num_reps, noise_amount)
 		line_num += 1
 		orig = float(line.strip())
 		with_noise = orig + gauss(GAUSSIAN_MEAN, noise_amount["sl1"])
-		lesioned.write("{}\n".format(with_noise))
+		lesioned.write("{:.6f}\n".format(with_noise))
 		if line_num == sl2_start - 1:
 			break
 
@@ -46,7 +50,7 @@ def lesion_assocs (patient_file, lesion_file, map_sizes, num_reps, noise_amount)
 		line_num += 1
 		orig = float(line.strip())
 		with_noise = orig + gauss(GAUSSIAN_MEAN, noise_amount["sl2"])
-		lesioned.write("{}\n".format(with_noise))
+		lesioned.write("{:.6f}\n".format(with_noise))
 		if line_num == l1l2_start - 1:
 			break
 
@@ -54,7 +58,7 @@ def lesion_assocs (patient_file, lesion_file, map_sizes, num_reps, noise_amount)
 		line_num += 1
 		orig = float(line.strip())
 		with_noise = orig + gauss(GAUSSIAN_MEAN, noise_amount["l1l2"])
-		lesioned.write("{}\n".format(with_noise))
+		lesioned.write("{:.6f}\n".format(with_noise))
 		if line_num == l1l2_inclusive_end:
 			break
 
@@ -62,12 +66,20 @@ if __name__ == "__main__":
 	map_sizes = {"l1": 20, "l2": 20, "sem": 20}
 	num_reps = {"l1": 208, "l2": 120, "sem": 81}
 
-	UTBA18_noise_amount = {'l1l2': 10, 'sl1': 10, 'sl2': 10}
+	UTBA18_noise_amount = {'l1l2': 0, 'sl1': 3, 'sl2': 3}
 	BUBA01_noise_amount = {'l1l2': 0, 'sl1': 0, 'sl2': 0}
 	UTBA20_noise_amount = {'l1l2': 0, 'sl1': 0, 'sl2': 0}
 	UTBA21_noise_amount = {'l1l2': 0, 'sl1': 0, 'sl2': 0}
 
-	lesion_assocs("../../UTBA18-simu", "UTBA18-lesioned-simu", map_sizes, num_reps, UTBA18_noise_amount)
+	UTBA18_overwrite = True
+	BUBA01_overwrite = False
+	UTBA20_overwrite = False
+	UTBA21_overwrite = False
+
+	lesion_assocs("../../UTBA18-simu", "../../UTBA18-lesioned-simu", map_sizes, num_reps, UTBA18_noise_amount, UTBA18_overwrite)
+	#lesion_assocs("../../BUBA01-simu", "../../BUBA01-lesioned-simu", map_sizes, num_reps, BUBA01_noise_amount, BUBA01_overwrite)
+	#lesion_assocs("../../UTBA20-simu", "../../UTBA20-lesioned-simu", map_sizes, num_reps, UTBA20_noise_amount, UTBA20_overwrite)
+	#lesion_assocs("../../UTBA21-simu", "../../UTBA21-lesioned-simu", map_sizes, num_reps, UTBA21_noise_amount, UTBA21_overwrite)
